% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddHelpText}
\alias{AddHelpText}
\title{Convert input text into HTML.}
\usage{
AddHelpText(content)
}
\arguments{
\item{content}{The text as a character string.}
}
\value{
HTML code snippet of the input string as a padded paragraph with custom font color.
}
\description{
Convert input text into HTML.
}
