% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddIntroductionTab}
\alias{AddIntroductionTab}
\title{Add Introduction tab in the body.}
\usage{
AddIntroductionTab(tabId, mainTitle, introText)
}
\arguments{
\item{tabId}{A unique id to identify the tab.}

\item{mainTitle}{The header of the tab.}

\item{introText}{The introduction of the tab.}
}
\value{
HTML code snippet to add an intro tabItem to main body.
}
\description{
Add Introduction tab in the body.
}
