% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddMultiTab}
\alias{AddMultiTab}
\title{Add multiple tabs in the body.}
\usage{
AddMultiTab(tabId, plotBox = TRUE, mainTitle, introText, headList,
  descList, plotList, descTitle = "Description", plotTitle = "Plot",
  descBoxWidth = 7, plotBoxWidth = 5)
}
\arguments{
\item{tabId}{A unique id to identify the tab.}

\item{plotBox}{The check if plot box should be added. (optional)}

\item{mainTitle}{The header of the tab.}

\item{introText}{The introduction of the tab.}

\item{headList}{The list of headings of the dimensions.}

\item{descList}{The list of descriptions of the plots.}

\item{plotList}{The list of plots of the tab.}

\item{descTitle}{The title of the description box. (optional)}

\item{plotTitle}{The title of the plot box. (optional)}

\item{descBoxWidth}{The width of the description box. (optional)}

\item{plotBoxWidth}{The width of the plot box. (optional)}
}
\value{
HTML code snippet to add a multiple tabItem to main body with 3 boxes:
* An introduction box
* A description box with multiple tabs
* A plot box with multiple tabs
}
\description{
Add multiple tabs in the body.
}
