% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddNetworkTab}
\alias{AddNetworkTab}
\title{Add a network tab in the body.}
\usage{
AddNetworkTab(tabId, mainTitle, introText, paramTitle = "Parameters",
  paramList, paramBoxWidth = 3, plotTitle = "Plot", plotList,
  plotBoxWidth = 9, plotHeight = "500px", sizeId, sizeList, colorId,
  colorList, opacityId)
}
\arguments{
\item{tabId}{A unique id to identify the tab.}

\item{mainTitle}{The header of the tab.}

\item{introText}{The introduction of the tab.}

\item{paramTitle}{The title of the input parameters box. (optional)}

\item{paramList}{The input widgets of the plot.}

\item{paramBoxWidth}{The width of the input parameters box. (optional)}

\item{plotTitle}{The title of the plot box. (optional)}

\item{plotList}{The plot of the tab.}

\item{plotBoxWidth}{The width of the plot box. (optional)}

\item{plotHeight}{The height of the plot in px. (optional)}

\item{sizeId}{A unique id to identify size input.}

\item{sizeList}{A named list of drop-down options for size of nodes. Values should be column names of input data.}

\item{colorId}{A unique id to identify color input.}

\item{colorList}{A named list of drop-down pptions for color of nodes. Values should be column names of input data.}

\item{opacityId}{A unique id to identify opacity input.}
}
\value{
HTML code snippet to add a network tabItem to main body with 3 boxes:
* An introduction box
* An input parameters box
* A plot box
}
\description{
Add a network tab in the body.
}
