% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddSimpleTab}
\alias{AddSimpleTab}
\title{Add a simple tab in the body.}
\usage{
AddSimpleTab(tabId, mainTitle, introText, descTitle = "Description",
  descList, descBoxWidth = 7, plotTitle = "Plot", plotList,
  plotBoxWidth = 5)
}
\arguments{
\item{tabId}{A unique id to identify the tab.}

\item{mainTitle}{The header of the tab.}

\item{introText}{The introduction of the tab.}

\item{descTitle}{The title of the description box. (optional)}

\item{descList}{The description of the plot.}

\item{descBoxWidth}{The width of the description box. (optional)}

\item{plotTitle}{The title of the plot box. (optional)}

\item{plotList}{The plot of the tab.}

\item{plotBoxWidth}{The width of the plot box. (optional)}
}
\value{
HTML code snippet to add a simple tabItem to main body with 3 boxes:
* An introduction box
* A description box
* A plot box
}
\description{
Add a simple tab in the body.
}
