% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{coord_radar}
\alias{coord_radar}
\title{Helper function for radar plot.}
\usage{
coord_radar(theta = "x", start = 0, direction = 1)
}
\arguments{
\item{theta}{TO-DO.}

\item{start}{TO-DO.}

\item{direction}{TO-DO.}
}
\value{
TO-DO.
}
\description{
Helper function for radar plot.
}
