% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createBar}
\alias{createBar}
\title{create a bar plot either single bar for each column or ...}
\usage{
createBar(userToken, data, uniform, color, xlabel, xlabs = NULL,
  single = FALSE, ylims = NULL, grouped = FALSE, grouplabs = NULL,
  groupsets = c(1), legendtitle = "LEGEND")
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}

\item{uniform}{The color of the bars, same for all bars or not.}

\item{color}{A single color variable or a vector of colors for the plot.}

\item{xlabel}{The x-axis label.}

\item{xlabs}{The x-axis tick labels.}

\item{single}{A boolean variable for single value for bar when TRUE, and mean column values when FALSE. (optional)}

\item{ylims}{The limits of y axis. (optional)}

\item{grouped}{A boolean variable for grouped bars. (optional)}

\item{grouplabs}{A character vector of labels of grouped bars. (optional)}

\item{groupsets}{An numeric vector showing which column belongs to which group. (optional)}

\item{legendtitle}{The title of the legend. (optional)}
}
\value{
a bar plot.
}
\description{
create a bar plot either single bar for each column or ...
}
