% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createBox}
\alias{createBox}
\title{Create box plot of self versus others.}
\usage{
createBox(userToken, dataSelf, dataOthers, dataOthersMean, xlims = NULL)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{dataSelf}{Data representing self value.}

\item{dataOthers}{Data representing others values.}

\item{dataOthersMean}{Data representing mean value among others.}

\item{xlims}{Limits of x axis. (optional)}
}
\description{
Create box plot of self versus others.
}
