% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createHist}
\alias{createHist}
\title{Create a histogram plot.}
\usage{
createHist(userToken, data, color, xlabel)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{A 1-dimensional array/ data frame of data.}

\item{color}{The color of histogram bars.}

\item{xlabel}{The x-axis label of the histogram.}
}
\value{
A histogram plot with vertical line to show user position.
}
\description{
Create a histogram plot.
}
