% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createIntroduction}
\alias{createIntroduction}
\title{Add introduction to the tab with multiple paragraphs.}
\usage{
createIntroduction(...)
}
\arguments{
\item{...}{Any number of characters strings to be converted into separate HTML paragraphs.}
}
\value{
HTML code snippet for single/multiple paragraphs.
}
\description{
Add introduction to the tab with multiple paragraphs.
}
