% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createNetwork}
\alias{createNetwork}
\title{Create a full network picture from network data.}
\usage{
createNetwork(userToken, inputNetwork, inputData, networkType,
  size = "uniform", color = "nNameNew", label, legend = FALSE,
  opacity = 1)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{inputNetwork}{Input network in the form of Adjacency Matrix.}

\item{inputData}{A data frame of attributes of nodes of inputNetwork.}

\item{networkType}{Network type to be created. Either "FULL" or "SMALL".}

\item{size}{Column used to define the size of nodes. (optional)}

\item{color}{COlumn used to define color of nodes. (optional)}

\item{label}{Column used to define label of nodes.}

\item{legend}{Boolean variable to show/not show the legend. (optional)}

\item{opacity}{opacity of nodes (optional)}
}
\value{
NetworkD3 dynamic graph.
}
\description{
Create a full network picture from network data.
}
