% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createPie}
\alias{createPie}
\title{Create Pie Chart.}
\usage{
createPie(userToken, data, colors)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}

\item{colors}{The vector representing colors of pie-slices.}
}
\value{
A pie chart with percentage labels.
}
\description{
Create Pie Chart.
}
