% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createQuotes}
\alias{createQuotes}
\title{Create quotes page.}
\usage{
createQuotes(userToken, data)
}
\arguments{
\item{userToken}{A user-specific password to show user quotes.}

\item{data}{input data of quotes.}
}
\value{
user defined quote blocks.
}
\description{
Create quotes page.
}
