% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createRadarComp}
\alias{createRadarComp}
\title{Create 2 radar plots for comparison.}
\usage{
createRadarComp(userToken, data, dataComp,
  legendColors = c(COLOR_DEFAULT_USER, COLOR_DEFAULT_PLOT), dim = " ",
  ylimMin = 0, ylimMax = NULL, legendLabels, varlabels)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}

\item{dataComp}{An input data frame for comparison.}

\item{legendColors}{A vector of both colors for radar plots. (optional)}

\item{dim}{Manually selected dimension (tab) highlighted when the relevant tab is selected. (optional)}

\item{ylimMin}{Lower limit of y-axis. (optional)}

\item{ylimMax}{Upper limit of y-axis. (optional)}

\item{legendLabels}{A character vector for legend values for two radars.}

\item{varlabels}{A character vector for axis labels.}
}
\value{
Two radar plots with different colors.
}
\description{
Create 2 radar plots for comparison.
}
