% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createScatter}
\alias{createScatter}
\title{Create Scatter Plot.}
\usage{
createScatter(userToken, data)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}
}
\value{
A scatter plot with highlighted user point.
}
\description{
Create Scatter Plot.
}
