% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{dataNetwork}
\alias{dataNetwork}
\title{Make network data usable in the form of edge list and node list.}
\usage{
dataNetwork(inputNetwork, inputData)
}
\arguments{
\item{inputNetwork}{A network as a two dimensional array with edges showns as matrix values of 1 and nodes as row/column names.}

\item{inputData}{The data of all the nodes in the above network.}
}
\value{
Edge list and Node list of the given data.
}
\description{
Make network data usable in the form of edge list and node list.
}
