% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egonets}
\alias{egonets}
\title{Sample ego networks for each user.}
\format{A list of 100 named elements where each element is a 2-D array.
\describe{
  \item{1}{Type = double[7x7], Value = 0,1,1,...}
  \item{2}{Type = double[4x4], Value = 0,1,1,...}
  ...
}}
\source{
TO-DO
}
\usage{
egonets
}
\description{
A named list of adjacency matrices, where names show user names and value is an adjacency matrix of ego network. 1 represent a connection, 0 represent no connection.
}
\keyword{datasets}
