% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{egonets_attrs}
\alias{egonets_attrs}
\title{Sample attributes for each user ego entwork.}
\format{A list of 100 named elements where each element is [n x 10] data frame.
\describe{
  for each data-frame
  \item{ID}{}
  \item{ID2}{}
  \item{cont1}{sample attribute}
  \item{cont2}{sample attribute}
  \item{cont3}{sample attribute}
  \item{cont4}{sample attribute}
  \item{cat1}{sample attribute}
  \item{cat2}{sample attribute}
  \item{cat3}{sample attribute}
  \item{cat4}{sample attribute}
}}
\usage{
egonets_attrs
}
\description{
A named list of data-frames, where names show user names and values are data frames of atrributes of users in a particular ego network.
}
\keyword{datasets}
