% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{formatLineData}
\alias{formatLineData}
\title{Format raw data into plot-able multi-line plot data.}
\usage{
formatLineData(userToken, data, xlength, dimnames = NULL)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}

\item{xlength}{Th length of x-axis/ The number of variables.}

\item{dimnames}{A chacter vector representing line labels in the legend. (optional)}
}
\value{
A data frame with first column as x-axis tick labels and one line per column for the rest of the columns.
}
\description{
Format raw data into plot-able multi-line plot data.
}
