% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_app.R
\name{list_app_templates}
\alias{list_app_templates}
\title{List available app templates}
\usage{
list_app_templates()
}
\description{
Lists all app templates available for scaffolding
}
\examples{
list_app_templates()
}
