% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_app.R
\name{list_mod_templates}
\alias{list_mod_templates}
\title{List available module templates}
\usage{
list_mod_templates()
}
\description{
Lists all module templates available for scaffolding
}
\examples{
list_mod_templates()
}
