% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{meltLineData}
\alias{meltLineData}
\title{Melt user-specific filtered data.}
\usage{
meltLineData(userToken, data)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}
}
\value{
A two-columns data frame with variable named as colx and column 2 containing values.
}
\description{
Melt user-specific filtered data.
}
