% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_app.R
\name{scaffold_app}
\alias{scaffold_app}
\title{Create boilerplate code from config file}
\usage{
scaffold_app(yaml_config_file)
}
\arguments{
\item{yaml_config_file}{String pointing to the path of a YAML file}
}
\description{
Creates the boilerplate code for a Shiny app as specified in
a YAML configuration file.
}
\examples{
# export config file from package to disk
yaml_config_file <- system.file("config.yml", package = "shinyscaffold")
yaml_config <- readLines(yaml_config_file)
writeLines(yaml_config, con = "/tmp/config.yml")
# -> modify /tmp/config.yml as needed

# scaffold app
setwd("/tmp")
scaffold_app("/tmp/config.yml")
}
