% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scaffold_app.R
\name{scaffold_config}
\alias{scaffold_config}
\title{Create scaffolding config file}
\usage{
scaffold_config(outfile = NULL)
}
\arguments{
\item{outfile}{String of a file name where the config file is exported to.
If left \code{NULL}, the config is printed to stdout.}
}
\description{
Creates YAML config file with the specification of the Shiny
app to scaffold. The file allows configuration of modules, styles and
other generic app features.
}
\examples{
# print config to screen
scaffold_config()

# write config to file
scaffold_config("/tmp/my-config.yml")
}
