% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/server.R
\name{server}
\alias{server}
\title{Shiny app server object.}
\usage{
server(input, output, session)
}
\arguments{
\item{input}{input provided by shiny app.}

\item{output}{output provided by shiny app.}

\item{session}{object to access session information.}
}
\description{
Shiny app server object.
}
