% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallnets}
\alias{smallnets}
\title{Sample small/individual networks for each user.}
\format{A list of 100 named elements where each element is a 2-D array.
\describe{
  \item{1}{Type = double[11x11], Value = 0,1,1,...}
  \item{2}{Type = double[2x2], Value = 0,1,1,...}
  ...
}}
\source{
TO-DO
}
\usage{
smallnets
}
\description{
A named list of adjacency matrices, where names show user names and value is an adjacency matrix of small/individual network. 1 represent a connection, 0 represent no connection.
}
\keyword{datasets}
