% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{smallnets_attrs}
\alias{smallnets_attrs}
\title{Sample attributes for each user small/individual network.}
\format{A list of 100 named elements where each element is [n x 5] data frame.
\describe{
  for each data-frame
  \item{ID}{}
  \item{cont1}{sample attribute}
  \item{cont2}{sample attribute}
  \item{cat1}{sample attribute}
  \item{cat2}{sample attribute}
}}
\usage{
smallnets_attrs
}
\description{
A named list of data-frames, where names show user names and values are data frames of atrributes of users in a particular small/individual network.
}
\keyword{datasets}
