% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{toynetwork}
\alias{toynetwork}
\title{Sample network for users' relations.}
\format{A symmetric 100 x 100 array.
\describe{
  \item{dim1}{IDs of all users.}
  \item{dim2}{IDs of all users.}
  ...
}}
\source{
TO-DO
}
\usage{
toynetwork
}
\description{
A network in the form of a symmetric matrix, 1 represent a connection, 0 represent no connection.
}
\keyword{datasets}
