% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ui.R
\docType{data}
\name{ui}
\alias{ui}
\title{Shiny app ui object.}
\format{An object of class \code{shiny.tag} of length 3.}
\usage{
ui
}
\description{
It is used to load in ui header, sidebar, and body components.
}
\keyword{datasets}
