% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{AddNetworkTab}
\alias{AddNetworkTab}
\title{Add a network tab in the body.}
\usage{
AddNetworkTab(tabId, mainTitle, introText, paramTitle = "Parameters",
  paramList, paramBoxWidth = 12, plotTitle = "Plot", plotList,
  plotBoxWidth = 12)
}
\arguments{
\item{tabId}{A unique id to identify the tab.}

\item{mainTitle}{The header of the tab.}

\item{introText}{The introduction of the tab.}

\item{paramTitle}{The title of the input parameters box. (optional)}

\item{paramList}{The input widgets of the plot.}

\item{paramBoxWidth}{The width of the input parameters box. (optional)}

\item{plotTitle}{The title of the plot box. (optional)}

\item{plotList}{The plot of the tab.}

\item{plotBoxWidth}{The width of the plot box. (optional)}
}
\value{
HTML code snippet to add a network tabItem to main body with 3 boxes:
* An introduction box
* An input parameters box
* A plot box
}
\description{
Add a network tab in the body.
}
