% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createNetwork}
\alias{createNetwork}
\title{Create a full network picture from network data.}
\usage{
createNetwork(userToken, nl, el, size = "uniform", color = "nNameNew",
  label, opacity = 1)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{nl}{Node list from input network and data.}

\item{el}{Edge list from input network and data.}

\item{size}{Column used to define the size of nodes. (optional)}

\item{color}{COlumn used to define color of nodes. (optional)}

\item{label}{Column used to define label of nodes. (optional)}

\item{opacity}{opacity of nodes (optional)}
}
\value{
NetworkD3 dynamic graph.
}
\description{
Create a full network picture from network data.
}
