% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{createRadar}
\alias{createRadar}
\title{Create a Radar Plot.}
\usage{
createRadar(userToken, data)
}
\arguments{
\item{userToken}{A user-specific password to show user position on the plot.}

\item{data}{An input data frame.}
}
\value{
A ggplot radar plot.
}
\description{
Create a Radar Plot.
}
